Moved LIDAR class to FrontElevator package -Removed Pneumatics package -cleaned up various files -added SmartDashboard information function to Robot
